Remove the `cli::clap` module altogether
